+++ client/components/detail/LocationModal.vue
