+++ client/layouts/index/IndexSearchLayout.vue
@@ -1,7 +1,7 @@
 <template>
   <ClientOnly>
-    <section class="SearchLayoutWrapper">    
-      <article class="Title"> 주변의 착한소비를 찾아보세요 </article>
+    <section class="SearchLayoutWrapper">
+      <article class="Title">주변의 착한소비를 찾아보세요</article>
       <SearchBar />
       <SearchArticles />
     </section>
@@ -14,10 +14,8 @@ import SearchArticles from '/components/index/search/SearchArticles.vue'
 </script>
 
 <style scoped lang="scss">
-
 $borderTest: 1px solid black;
 .SearchLayoutWrapper {
-
   font-family: 'Pretendard';
   display: flex;
   flex-direction: column;
@@ -26,13 +24,13 @@ $borderTest: 1px solid black;
   min-height: 994px;
   padding-bottom: 80px;
 
-  background-color: #F9FAFB;
-  
-  .Title {           
+  background-color: #f9fafb;
+
+  .Title {
     margin-top: 80px;
     font-size: 28px;
     font-weight: 700;
     line-height: 150%;
-  } 
+  }
 }
 </style>
