+++ client/pages/index.vue
@@ -1,11 +1,11 @@
 <template>
   <ClientOnly>
     <!-- 진원 -->
-    <HeaderLayout />         
+    <HeaderLayout />
     <LandingLayout />
 
     <!-- 준하 -->
-    <SearchLayout />
+    <IndexSearchLayout />
     <HowToLayout />
 
     <!-- 진원 -->
@@ -14,9 +14,9 @@
 </template>
 
 <script setup lang="ts">
-import HeaderLayout from '/layouts/HeaderLayout.vue';
+import HeaderLayout from '/layouts/HeaderLayout.vue'
 import LandingLayout from '/layouts/index/LandingLayout.vue'
-import SearchLayout from '/layouts/index/SearchLayout.vue'
+import IndexSearchLayout from '/layouts/index/IndexSearchLayout.vue'
 import HowToLayout from '/layouts/index/HowToLayout.vue'
 import FooterLayout from '/layouts/FooterLayout.vue'
 </script>
